// frontend/src/app/api/model/alumno.ts
/**
 * API oPROsita
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Usuario } from './usuario';


export interface Alumno extends Usuario { 
    tipo?: Alumno.TipoEnum;
}
export namespace Alumno {
    export const TipoEnum = {
        Alumno: 'alumno'
    } as const;
    export type TipoEnum = typeof TipoEnum[keyof typeof TipoEnum];
}